requirements.txt
flask>=2.3
pyjwt
werkzeug
pytest
flask-sqlalchemy>=3.0
